[FIX] <pro>Table: Fix the problem that editing cannot be cancelled when editing in-line.
