+++ CHANGELOG.zh-CN.md
@@ -25,6 +25,7 @@ timeline: true
 - 🐞 `<pro>Pagination`: 修复最大分页没有设置时，分页选项有超过默认最大100分页无法选择的问题。
 - 🐞 `<pro>IconPicker`: 修复当点击分页按钮并按钮成禁用状态时，弹窗无法关闭的问题。
 - 🐞 `<pro>Table`: 修复在孔雀蓝主题下，当Column的editor为function的时候，进入编辑态点击退出后会导致下次进入编辑态组件宽度不正确的问题
+- 🐞 `<pro>Table`: 修复行内编辑时无法取消编辑的问题。
 
 ## 1.4.1
